Update RLTK to from 0.5.x to 0.6.0

+++ hellorust/src/monster_ai_system.rs
@@ -90,8 +90,8 @@ impl<'a> System<'a> for MonsterAI {
                     if path.success && path.steps.len() > 1 {
                         let mut idx = map.xy_idx(pos.x, pos.y);
                         map.blocked_tiles[idx] = false;
-                        pos.x = path.steps[1] % map.width;
-                        pos.y = path.steps[1] / map.width;
+                        pos.x = path.steps[1] as i32 % map.width;
+                        pos.y = path.steps[1] as i32 / map.width;
                         idx = map.xy_idx(pos.x, pos.y);
                         map.blocked_tiles[idx] = true;
                         viewshed.dirty = true;
